refactoring ListCategory
