+++ frontend/src/components/Overmoney/Overmoney.tsx
@@ -45,53 +45,8 @@ const cards = [
 
 ] as ICard[];
 
-const constlistItems = [
-    {
-        id: 1,
-        name: 'Категория 1',
-        type: "INCOME",
-        keywords: ['Ключевое слово 1', 'Ключевое слово 2', 'Ключевое слово 3'],
-    },
-    {
-        id: 3,
-        name: 'Категория 3',
-        type: "INCOME",
-        keywords: ['Ключевое слово 1', 'Ключевое слово 2', 'Ключевое слово 3'],
-    },
-    {
-        id: 2,
-        name: 'Категория 2',
-        type: "EXPENSE",
-        keywords: ['Ключевое слово 1', 'Ключевое слово 2', 'Ключевое слово 3'],
-    },
-    {
-        id: 4,
-        name: 'Категория 4',
-        type: "EXPENSE",
-        keywords: ['Ключевое слово 1', 'Ключевое слово 2', 'Ключевое слово 3'],
-    },
 
-] as IListItem[];
 const Overmoney: FC = () => {
-
-    const [listItems, setListItems] = React.useState<IListItem[]>(constlistItems);
-
-    const [showModalAddCategory, setShowModalAddCategory] = useState<boolean>(false);
-    const handleCloseModalAddCategory = () => setShowModalAddCategory(false);
-    const handleShowModalAddCategory = () => setShowModalAddCategory(true);
-
-    const handleSubmitAddCategory = useCallback((formData: IListItem) => {
-        //вызов API добавления категории 
-        setListItems([
-            ...listItems,
-            {
-                name: formData.name,
-                type: formData.type,
-                keywords: [formData.name]
-            }
-        ]);
-        handleCloseModalAddCategory();
-    }, [listItems]);
 
     return (
         <>
@@ -101,15 +56,11 @@ const Overmoney: FC = () => {
                         <GridCards cards={cards} />
                     </Col>
                     <Col sm={4}>
-                        <ListCategories items={listItems} handleClickAddCategory={handleShowModalAddCategory} />
+                        <ListCategories />
                     </Col>
                 </Row>
             </Container>
-            <PopupAddCategory 
-                showModalAddCategory={showModalAddCategory} 
-                handleCloseModalAddCategory={handleCloseModalAddCategory} 
-                handleButtonSubmit={handleSubmitAddCategory}
-            />
+            
         </>
     );
 };
